Created Id Tracking for Solves
In order to allow for dynamic statistics tracking
I modified the code to keep track of every solve
time and solve id. This will also allow for deletion
and recalculation of all global statistics.

This was done by converting the time array, which
was a basic array, into an array of object, which
keeps track of the id and solveTime.