Added Continual Mean of Times
The changes on this commit allow for the mean to be updated from
the sum of all the times. This allows a larger precision than if
it was just a mean that was updated with the new times.

It also sets up the procedure to allow for the other statistics
to be updated without a large amount of overhead being added